fix: yt-dlp search duration parsing, unblock Spotify track downloads
YtDlpSearchResult.duration changed from Long to Double because yt-dlp
returns duration as a float (e.g. 396.0). The previous Long type caused
kotlinx.serialization to fail, which silently dropped all search results
and caused Spotify-sourced tracks to fail downloading with "unmatched".

+++ data/download/src/main/kotlin/com/stash/data/download/matching/MatchScorer.kt
@@ -68,7 +68,7 @@ class MatchScorer @Inject constructor(
         return results.map { result ->
             val titleScore = computeTitleScore(targetTitle, result.title)
             val artistScore = computeArtistScore(targetArtist, result.uploader)
-            val durationScore = computeDurationScore(targetDurationMs, result.duration)
+            val durationScore = computeDurationScore(targetDurationMs, result.duration.toLong())
             val popularityScore = computePopularityScore(result.viewCount, maxViewCount)
             val penalty = computePenalty(targetTitle, result.title)
             val topicBonus = computeTopicBonus(result.uploader, result.channel)
@@ -88,7 +88,7 @@ class MatchScorer @Inject constructor(
                 videoId = result.id,
                 title = result.title,
                 uploader = result.uploader,
-                durationSeconds = result.duration,
+                durationSeconds = result.duration.toLong(),
                 viewCount = result.viewCount,
                 matchScore = finalScore,
             )
